Added Replay Buffer
Performance is way worse now that we are not sampling the full buffer ... we will need to hyperparameter tune from scratch.
While doing this, I also implemented (atm not used) soft Q network updates.

next is double DQN, as we are atm observing diverging (but slowly, so thats good) Q values