rewrite hyperparameter search procedure
Instead of creating a new simulation for each permutation of hyperparameters, an instance of sim controller now takes a list of hyperparameter permutations and iterates through them by resetting simulations.

Simplified hyperparameter indexing scheme so that only 1 "umbrella" database is created for one configuration file (instead of 1 database per hyperparameter permutation). This should be the expected behaviour that is consistent with all other simulation types.

A new "hyperparameters" table is created to make it easier to search permutation by index